+++ Fabric/src/main/java/ca/lukegrahamlandry/lib/event/fabric/WrapperLibModInitializer.java
@@ -11,6 +11,7 @@ package ca.lukegrahamlandry.lib.event.fabric;
 
 import ca.lukegrahamlandry.lib.base.Available;
 import ca.lukegrahamlandry.lib.base.event.EventWrapper;
+import ca.lukegrahamlandry.lib.config.EventCallbacks;
 import ca.lukegrahamlandry.lib.keybind.KeybindTickCallbacks;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
@@ -35,5 +36,8 @@ public class WrapperLibModInitializer implements ModInitializer {
             ServerTickEvents.END_WORLD_TICK.register(
                     world -> world.players().forEach(
                             p -> KeybindTickCallbacks.onServerPlayerTick(p)));  // if you use method reference it tries to class load LocalPlayer, idk man
+
+        if (Available.CONFIG.get())
+            ServerLifecycleEvents.START_DATA_PACK_RELOAD.register((server, resourceManager) -> EventCallbacks.onReloadCommand());
     }
 }